create recivedEntity

+++ src/main/java/com/example/spread/domain/repository/SpreadRepository.java
@@ -3,6 +3,6 @@ import com.example.spread.domain.entity.SpreadEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 
-public interface SpreadRepository extends JpaRepository<SpreadEntity, Long>{
+public interface SpreadRepository extends JpaRepository<SpreadEntity, Integer>{
     // No need implementation, Just one interface, and you have CRUD,
 }
